server/requirements.txt
fastapi
uvicorn
python-socketio
tensorflow
pillow
requests
numpy
